Update ignores
git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/rgl/pkg/rgl@1364 79a1b648-8b68-254f-b20b-c0b50f5137bf

+++ src/glgui.cpp
@@ -99,7 +99,7 @@ double GLBitmapFont::height() {
 
 bool GLBitmapFont::valid(const char* text) {
   for (int i=0; text[i]; i++)
-    if ((int)text[i] < (int)firstGlyph || text[i] - firstGlyph >= (int)nglyph)
+    if ((int)text[i] < (int)firstGlyph || (int)text[i] - (int)firstGlyph >= (int)nglyph)
       return false;
   return true;
 }
